Add modify(data:)
This adds and implements `open func modify(data: [(String, Any)]) -> [(String, Any)]` to the `PostgresStORM` class.
The intent of this addition is to allow base classes to optionally change the data sent to the database. For example, I am using it to properly handle the `Date` class as follows:
```swift
override func modify(data: [(String, Any)]) -> [(String, Any)] {
	var dat = data
	for (i, v) in dat.enumerated() {
		if let d = v.1 as? Date {
			dat[i] = (v.0, d.timestamptz)
		}
	}
	
	return dat
}
```
Where `timestamptz` returns the date formatted as a PostgreSQL `timestamp with time zone` string.